fix: change BIGINT to INT for user_id

+++ migrations/003_refresh_tokens.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE TABLE refresh_tokens (
     id         BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    user_id    BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id    INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     token_hash TEXT NOT NULL UNIQUE,
     expires_at TIMESTAMPTZ NOT NULL,
     revoked_at TIMESTAMPTZ,
